scripts/requirements.txt
langchain_core==0.3.37
python-dotenv==1.0.1
Requests==2.32.3
scrapegraphai==1.39.0
torch==2.6.0
transformers==4.49.0
